First commit of a script to transmit data
A script to transmit data.

It has the ability to not run itself twice, thanks to PID file.

Small modification to transmit_file.py and comm_handler.py.

Comm_handler.py now does not print anything at all.
transmit_file has the ability to disable verbose prints, which is used
in the script.

Script has several hardcodes in it at the moment.